requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
django-heroku==0.3.1
django-parler==2.3
django-parler-rest==2.2
django-phonenumber-field==6.3.0
django-sesame==3.1
djangorestframework==3.13.1
fonttools==4.34.4
geographiclib==1.52
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
kiwisolver==1.4.4
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.53
Pillow==9.2.0
pluggy==1.0.0
protobuf==4.21.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.10
whitenoise==6.2.0
